requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
sentence-transformers==2.3.1
faiss-cpu>=1.8.0
numpy>=1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1
torch==2.2.0 --extra-index-url https://download.pytorch.org/whl/cpu
